--- manifest.mf
@@ -1,7 +1,0 @@
-Manifest-Version: 1.0
-AutoUpdate-Show-In-Client: true
-OpenIDE-Module: org.urmiauniversity.it.mst.cpm
-OpenIDE-Module-Localizing-Bundle: org/urmiauniversity/it/mst/cpm/Bundle.properties
-OpenIDE-Module-Requires: org.openide.windows.WindowManager
-OpenIDE-Module-Specification-Version: 1.0
-
