+++ modules/CPMplugin/README.md
@@ -1,0 +1,4 @@
+## CPM plugin
+
+This README supports Markdown, see [syntax](https://help.github.com/articles/markdown-basics/)
+
