+++ modules/CPMplugin/src/main/java/com/plugin/Bundle.properties
@@ -1,0 +1,17 @@
+OpenIDE-Module-Display-Category=Tools
+OpenIDE-Module-Long-Description=\
+    Simple algorithm for detecting overlapping communities based on node centeric method. \
+    The main reference is: G. \
+    Palla, I. \
+    Der\u00e9nyi, I. \
+    Farkas, and T. \
+    Vicsek. \
+    Uncovering the overlapping community structure of complex networks in nature and society. \
+    Nature, 435:814\u2013818, 2005. \
+    DOI: 10.1038/nature0360733, 78
+OpenIDE-Module-Name=Clique Percolation Method
+OpenIDE-Module-Short-Description=Clique Percolation Method implementation
+CPM_MSTTopComponent.jTextField1.text=jTextField1
+CPM_MSTTopComponent.jLabel1.text=Size of k:
+CPM_MSTTopComponent.jLabel1.toolTipText=K is the clique size which is the new clluster will be based on it.
+CPM_MSTTopComponent.jButton1.text=Start
