--- nbproject/project.properties
@@ -1,5 +1,0 @@
-javac.source=1.6
-javac.compilerargs=-Xlint -Xlint:-serial
-nbm.homepage=http://cse.urmia.ac.ir/~mst
-nbm.module.author=Mir Saman Tajbakhsh
-nbm.needs.restart=true
