--- nbproject/suite.properties
@@ -1,1 +1,0 @@
-suite.dir=${basedir}/..
